:recycle: Split into other classes in files
